+++ projects/packages/search/src/dashboard/components/donut-meter-container/index.jsx
@@ -11,33 +11,63 @@ import React from 'react';
 
 import './style.scss';
 
-// Format numbers with separators.
-const formatNumberWithSeparators = x => {
+const localizedUnlimited = __( 'Unlimited', 'jetpack-search-pkg' );
+
+// Format numbers with separators or convert to `Unlimited`.
+export const formatNumber = x => {
+	if ( x > 1e18 ) {
+		return localizedUnlimited;
+	}
+
 	return numberFormat( x );
+};
+
+const usageInfoMessage = ( current, limit ) => {
+	const isUnlimitedRequests = limit > 1e18;
+
+	// Standard case of current/limit.
+	if ( ! isUnlimitedRequests ) {
+		return `${ formatNumber( current ) } / ${ formatNumber( limit ) }`;
+	}
+
+	// Special case for "unlimited" requests. API returning 64-bit PHP_INT_MAX.
+	// Return "Unlimited" if the current count is zero.
+	if ( current === 0 ) {
+		return localizedUnlimited;
+	}
+
+	// We have a request count so include it in the info string.
+	// ie: "123/Unlimited"
+	return `${ formatNumber( current ) } / ${ localizedUnlimited }`;
 };
 
 /**
  * Returns a DonutMeterContainer describing resource usage.
  *
- * @param {object}prop - props to show usage info.
- * @param {number}prop.current - totalCount to the DonutMeter.
- * @param {number}prop.limit - segmentCount to the DonutMeter.
- * @param {string}prop.title - title to the DonutMeter.
- * @param {object}prop.tooltip - tooltip data
- * @returns {React.Component} DonutMeterContainer component.
+ * @param {object} props - Props
+ * @param {number} props.current - totalCount to the DonutMeter
+ * @param {number} props.limit - segmentCount to the DonutMeter
+ * @param {string} props.title - title to the DonutMeter
+ * @param {Function} props.iconClickedCallback - handler for click on "info" icon
+ * @param {Function} props.linkClickedCallback - handler for click on "details" link
+ * @param {object} props.tooltip - tooltip data
+ * @returns {React.Component} DonutMeterContainer component
  */
-const DonutMeterContainer = ( { current = 0, limit = 1, title, tooltip } ) => {
-	// TODO: Remove local callback in favour of props.
-	const tempCallback = () => {
-		// eslint-disable-next-line no-console
-		console.log( 'higher level callback...' );
-	};
-
-	const usageInfo =
-		formatNumberWithSeparators( current ) + '/' + formatNumberWithSeparators( limit );
+const DonutMeterContainer = ( {
+	current = 0,
+	limit = 1,
+	title,
+	tooltip,
+	iconClickedCallback,
+	linkClickedCallback,
+} ) => {
+	const isUnlimitedRequests = limit > 1e18;
+	const displayCurrent = isUnlimitedRequests ? 1 : current;
+	const displayLimit = isUnlimitedRequests ? 1 : limit;
+	const usageInfo = usageInfoMessage( current, limit );
 
 	const tooltipArgs = {
-		shadowAnchor: true,
+		popoverAnchorStyle: 'wrapper',
 		title: tooltip.title,
 		placement: 'top',
 		forceShow: tooltip.forceShow,
@@ -47,7 +77,11 @@ const DonutMeterContainer = ( { current = 0, limit = 1, title, tooltip } ) => {
 		<ThemeProvider>
 			<div className="donut-meter-container">
 				<div className="donut-meter-wrapper">
-					<DonutMeter segmentCount={ current } totalCount={ limit } />
+					<DonutMeter
+						segmentCount={ displayCurrent }
+						totalCount={ displayLimit }
+						useAdaptiveColors={ ! isUnlimitedRequests }
+					/>
 					<div className="upgrade-tooltip-shadow-anchor">
 						<IconTooltip { ...tooltipArgs }>
 							<>
@@ -61,8 +95,11 @@ const DonutMeterContainer = ( { current = 0, limit = 1, title, tooltip } ) => {
 					</div>
 				</div>
 				<div className="donut-info-wrapper">
-					<InfoPrimary localizedMessage={ title } iconClickedCallback={ tempCallback } />
-					<InfoSecondary localizedMessage={ usageInfo } linkClickedCallback={ tempCallback } />
+					<InfoPrimary localizedMessage={ title } iconClickedCallback={ iconClickedCallback } />
+					<InfoSecondary
+						localizedMessage={ usageInfo }
+						linkClickedCallback={ linkClickedCallback }
+					/>
 				</div>
 			</div>
 		</ThemeProvider>