+++ projects/plugins/jetpack/_inc/lib/class.core-rest-api-endpoints.php
@@ -2795,15 +2795,6 @@ class Jetpack_Core_Json_Api_Endpoints {
 				'jp_group'          => 'settings',
 			),
 
-			// Apps card on dashboard.
-			'dismiss_dash_app_card'                => array(
-				'description'       => '',
-				'type'              => 'boolean',
-				'default'           => 0,
-				'validate_callback' => __CLASS__ . '::validate_boolean',
-				'jp_group'          => 'settings',
-			),
-
 			// Empty stats card dismiss.
 			'dismiss_empty_stats_card'             => array(
 				'description'       => '',