+++ projects/plugins/jetpack/json-endpoints/class.wpcom-json-api-site-settings-v1-4-endpoint.php
@@ -74,6 +74,7 @@ new WPCOM_JSON_API_Site_Settings_V1_4_Endpoint(
 			'disallowed_keys'                         => '(string) Words or phrases that mark comment spam, one per line',
 			'lang_id'                                 => '(int) ID for language blog is written in',
 			'locale'                                  => '(string) locale code for language blog is written in',
+			'site_vertical_id'                        => '(string) The site vertical ID',
 			'wga'                                     => '(array) Google Analytics Settings',
 			'jetpack_cloudflare_analytics'            => '(array) Cloudflare Analytics Settings',
 			'disabled_likes'                          => '(bool) Are likes globally disabled (they can still be turned on per post)?',
@@ -111,6 +112,9 @@ new WPCOM_JSON_API_Site_Settings_V1_4_Endpoint(
 			'posts_per_page'                          => '(int) Number of posts to show on blog pages',
 			'posts_per_rss'                           => '(int) Number of posts to show in the RSS feed',
 			'rss_use_excerpt'                         => '(bool) Whether the RSS feed will use post excerpts',
+			'wpcom_publish_posts_with_markdown'       => '(bool) Whether markdown is enabled for posts',
+			'wpcom_publish_comments_with_markdown'    => '(bool) Whether markdown is enabled for comments',
+			'launchpad_screen'                        => '(string) Whether or not launchpad is presented and what size it will be',
 		),
 
 		'response_format' => array(